+++ Procfile
@@ -1,1 +1,2 @@
-web: python push.py
+web: python push_first.py
+web: python push_second.py
